# MyApi/openapi/v1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "MyApi | v1",
    "version": "1.0.0"
  },
  "paths": {
    "/weatherforecast": {
      "get": {
        "tags": [
          "WeatherHandlers"
        ],
        "summary": "Get weather forecast",
        "description": "Get weather forecast data for the next 5 days.",
        "operationId": "GetWeatherForecast",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeatherForecastDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/weatherforecast/generate": {
      "get": {
        "tags": [
          "WeatherHandlers"
        ],
        "summary": "Generate a plausible multi-day forecast using Semantic Kernel",
        "operationId": "GenerateWeatherForecast",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/weatherforecast/analyze": {
      "post": {
        "tags": [
          "MyApi"
        ],
        "operationId": "AnalyzeWeather",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/WeatherForecastDto"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/weatheranalysis/latest": {
      "get": {
        "tags": [
          "MyApi"
        ],
        "summary": "Get latest weather analysis",
        "operationId": "GetLatestWeatherAnalysis",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeatherAnalysisDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/weatherforecast/study": {
      "post": {
        "tags": [
          "WeatherHandlers"
        ],
        "summary": "Interactive weather study using Semantic Kernel",
        "operationId": "WeatherStudy",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WeatherStudyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/study/start": {
      "post": {
        "tags": [
          "MyApi"
        ],
        "summary": "Start a study session",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartStudyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/study/send": {
      "post": {
        "tags": [
          "MyApi"
        ],
        "summary": "Send a message and get assistant reply",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/goals": {
      "post": {
        "tags": [
          "MyApi"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGoalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/concepts/{goalId}": {
      "post": {
        "tags": [
          "MyApi"
        ],
        "parameters": [
          {
            "name": "goalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/quiz/{conceptId}": {
      "post": {
        "tags": [
          "MyApi"
        ],
        "parameters": [
          {
            "name": "conceptId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreateGoalRequest": {
        "required": [
          "description"
        ],
        "type": "object",
        "properties": {
          "description": {
            "type": "string"
          }
        }
      },
      "SendMessageRequest": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "format": "uuid"
          },
          "message": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "StartStudyRequest": {
        "type": "object",
        "properties": {
          "topic": {
            "type": "string",
            "nullable": true
          },
          "level": {
            "type": "string",
            "nullable": true
          },
          "exam": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "WeatherAnalysisDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "analysis": {
            "type": "string",
            "nullable": true
          },
          "userPrompt": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "WeatherForecastDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "nullable": true
          },
          "temperatureC": {
            "type": "integer",
            "format": "int32"
          },
          "temperatureF": {
            "type": "integer",
            "format": "int32"
          },
          "summary": {
            "type": "string",
            "nullable": true
          },
          "icon": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "WeatherStudyRequest": {
        "type": "object",
        "properties": {
          "forecasts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WeatherForecastDto"
            },
            "nullable": true
          },
          "userPrompt": {
            "type": "string",
            "nullable": true
          },
          "purpose": {
            "type": "string",
            "nullable": true
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "WeatherHandlers"
    },
    {
      "name": "MyApi"
    }
  ]
}