# Shared/Runtime/WeatherForecastDto.cs
namespace Dtos
{
    public sealed class WeatherForecastDto
    {
        public string Date { get; set; }
        public int TemperatureC { get; set; }
        public int TemperatureF { get; set; }
        public string Summary { get; set; }
        public string Icon { get; set; }
    }
}